requirements.txt
requests
PyJWT
cryptography
python-dotenv
openpyxl